Only run `getStories` tests on MacOS
which is unfortunate... 😞.

mock-fs mocks out the filesystem. It works great on all platforms for anything using the `fs`
module. However, it only seems to mock `require` calls effectively on MacOS. These tests pass
locally, but fail on CI on Ubuntu and Windows.

Possibly related to https://github.com/tschaub/mock-fs/issues/312.

An alternative would be to use Jest's module mocking to directly mock the `glob.sync` call and
calls to `require` for the individual story files. Doing so would likely work on all platforms,
but wouldn't be as good of a test since it mocks implementation details about `getStories`.